requirements.txt
streamlit
google-generativeai # Ensure to specify the latest version or the one you're using
python-dotenv